moved trasform to message-event.ts
